+++ src/Simulation.spec.ts
@@ -4,8 +4,9 @@ import { Simulation } from "./Simulation";
 import { Cell } from "./cultures/Cell";
 import { CellFactory } from "./cultures/CellFactory";
 import { TerrainFactory } from "./terrain/TerrainFactory";
-import { generateIds } from "./util/id/generator";
 import { Spot } from "./terrain/Spot";
+
+const generateIds = (n: number) => Array.from({ length: n }, (_, i) => `c${i}`);
 
 describe("E2E simulation test", () => {
   let terrain: Terrain, cells: Cell[], simulation: Simulation;
@@ -14,7 +15,7 @@ describe("E2E simulation test", () => {
   describe("when terrain and cultures are being created", () => {
     beforeEach(() => {
       terrain = TerrainFactory.createTerrain({ width: 3, height: 3, resourcesPerSpot: 1 });
-      cells = generateIds(3).map((id) => CellFactory.createCell(terrain, { id }));
+      cells = generateIds(3).map((id) => CellFactory.createCell(id, terrain));
       simulation = new Simulation(terrain, cells);
     });
 
@@ -30,7 +31,7 @@ describe("E2E simulation test", () => {
   describe("when resources are available for the cell", () => {
     beforeEach(() => {
       terrain = TerrainFactory.createTerrain({ width: 1, height: 1, resourcesPerSpot: 1 });
-      cells = [CellFactory.createCell(terrain, { id: "c0" })];
+      cells = [CellFactory.createCell("c0", terrain)];
       simulation = new Simulation(terrain, cells);
       simulation.step();
     });
@@ -43,7 +44,7 @@ describe("E2E simulation test", () => {
   describe("when resources are not available for the cell", () => {
     beforeEach(() => {
       terrain = TerrainFactory.createTerrain({ width: 1, height: 2, resourcesPerSpot: 1 });
-      cells = [CellFactory.createCell(terrain, { id: "c0" })];
+      cells = [CellFactory.createCell("c0", terrain)];
       simulation = new Simulation(terrain, cells);
 
       simulation.step();
@@ -60,7 +61,7 @@ describe("E2E simulation test", () => {
   describe("when there is no free spot for the cell to migrate", () => {
     beforeEach(() => {
       terrain = TerrainFactory.createTerrain({ width: 2, height: 2, resourcesPerSpot: 0 });
-      cells = generateIds(4).map((id) => CellFactory.createCell(terrain, { id }));
+      cells = generateIds(4).map((id) => CellFactory.createCell(id, terrain));
       simulation = new Simulation(terrain, cells);
 
       initialSpot = terrain.getSpot("c0");
@@ -76,7 +77,7 @@ describe("E2E simulation test", () => {
   describe("when a cell consume resources", () => {
     beforeEach(() => {
       terrain = TerrainFactory.createTerrain({ width: 1, height: 1, resourcesPerSpot: 1 });
-      cells = [CellFactory.createCell(terrain, { id: "c0" })];
+      cells = [CellFactory.createCell("c0", terrain)];
       simulation = new Simulation(terrain, cells);
       simulation.step();
     });
@@ -89,7 +90,7 @@ describe("E2E simulation test", () => {
   describe("when a cell has energy", () => {
     beforeEach(() => {
       terrain = TerrainFactory.createTerrain({ width: 1, height: 3, resourcesPerSpot: 0 });
-      cells = [CellFactory.createCell(terrain, { id: "c0" })];
+      cells = [CellFactory.createCell("c0", terrain)];
       simulation = new Simulation(terrain, cells);
 
       simulation.step();
@@ -109,7 +110,7 @@ describe("E2E simulation test", () => {
   describe("when cell has no energy", () => {
     beforeEach(() => {
       terrain = TerrainFactory.createTerrain({ width: 1, height: 2, resourcesPerSpot: 0 });
-      cells = [CellFactory.createCell(terrain, { id: "c0", initialEnergy: 0.2 })]; // cell will consume 0.2 for inner processes, so no energy for migration
+      cells = [CellFactory.createCell("c0", terrain, { initialEnergy: 0.2 })]; // cell will consume 0.2 for inner processes, so no energy for migration
       simulation = new Simulation(terrain, cells);
 
       initialSpot = terrain.getSpot("c0");
@@ -127,7 +128,7 @@ describe("E2E simulation test", () => {
   describe("when cell has too much energy", () => {
     beforeEach(() => {
       terrain = TerrainFactory.createTerrain({ width: 1, height: 1, resourcesPerSpot: 1 });
-      cells = [CellFactory.createCell(terrain, { id: "c0", initialEnergy: 50 })];
+      cells = [CellFactory.createCell("c0", terrain, { initialEnergy: 50 })];
       simulation = new Simulation(terrain, cells);
       simulation.step(); // each step consumes 0.7
       simulation.step();
@@ -143,7 +144,7 @@ describe("E2E simulation test", () => {
   describe("when cell has enough energy to divide", () => {
     beforeEach(() => {
       terrain = TerrainFactory.createTerrain({ width: 1, height: 2, resourcesPerSpot: 10 });
-      cells = [CellFactory.createCell(terrain, { id: "c0", initialEnergy: 1 })];
+      cells = [CellFactory.createCell("c0", terrain, { initialEnergy: 1 })];
       simulation = new Simulation(terrain, cells);
 
       for (let i = 0; i < 18; i++) {
@@ -160,7 +161,7 @@ describe("E2E simulation test", () => {
   describe("when cell has not enough energy to divide", () => {
     beforeEach(() => {
       terrain = TerrainFactory.createTerrain({ width: 1, height: 2, resourcesPerSpot: 0 });
-      cells = [CellFactory.createCell(terrain, { id: "c0", initialEnergy: 1, initialSize: 8 })];
+      cells = [CellFactory.createCell("c0", terrain, { initialEnergy: 1, initialSize: 8 })];
       simulation = new Simulation(terrain, cells);
 
       for (let i = 0; i < 18; i++) {