+++ src/cultures/Cell.ts
@@ -3,30 +3,52 @@ import { TerrainFacade } from "../terrain/TerrainFacade";
 import { EnergyStore, NotEnoughEnergyError } from "../energy/EnergyStore";
 import { randomStr } from "../util/random";
 
-const INITIAL_SIZE = 1;
-const INITIAL_ENERGY = 5;
-const RESOURCE_INTAKE = 1;
-const INNER_PROCESS_ENERGY_RATE = 0.2;
-const GROW_PROCESS_ENERGY_RATE = 0.5;
-const MIGRATION_PROCESS_ENERGY_RATE = 0.2;
-const DIVISION_PROCESS_ENERGY_RATE = 5;
 const MINIMAL_SIZE_TO_DIVIDE = 8;
 
 export type DivisionCallback = (newCell: Cell) => void;
 
+export interface CellParameters {
+  initialSize?: number;
+  initialEnergy?: number;
+  migrationEnergyCost?: number;
+  divisionEnergyCost?: number;
+  growEnergyCost?: number;
+  innerProcessEnergyCost?: number;
+  resourceConsumptionRate?: number;
+  growRate?: number;
+}
+
+type FulfilledCellParameters = Required<CellParameters>;
+
+const DEFAULT_CELL_PARAMETERS: FulfilledCellParameters = {
+  initialSize: 1,
+  initialEnergy: 5,
+  migrationEnergyCost: 0.2,
+  divisionEnergyCost: 5,
+  growEnergyCost: 0.5,
+  innerProcessEnergyCost: 0.2,
+  resourceConsumptionRate: 1,
+  growRate: 0.5,
+};
+
+const ensureParametersFullfilled = (parameters: CellParameters): FulfilledCellParameters => {
+  return { ...DEFAULT_CELL_PARAMETERS, ...parameters };
+};
+
 export class Cell {
   private readonly energyStore: EnergyStore;
+  private readonly parameters: FulfilledCellParameters;
   private size: number;
   private children: Cell[] = [];
 
   constructor(
     public readonly id: string,
     private readonly terrain: TerrainFacade,
-    initialEnergy: number = INITIAL_ENERGY,
-    initialSize: number = INITIAL_SIZE,
+    config: CellParameters = DEFAULT_CELL_PARAMETERS,
   ) {
-    this.energyStore = new EnergyStore(initialEnergy);
-    this.size = initialSize;
+    this.parameters = ensureParametersFullfilled(config);
+    this.energyStore = new EnergyStore(this.parameters.initialEnergy);
+    this.size = this.parameters.initialSize;
   }
 
   get energy(): number {
@@ -38,7 +60,7 @@ export class Cell {
   }
 
   private get canDivide(): boolean {
-    return this.size >= MINIMAL_SIZE_TO_DIVIDE && this.energyStore.currentValue >= DIVISION_PROCESS_ENERGY_RATE;
+    return this.size >= MINIMAL_SIZE_TO_DIVIDE && this.energyStore.currentValue >= this.parameters.divisionEnergyCost;
   }
 
   public step() {
@@ -50,7 +72,7 @@ export class Cell {
 
   private executeInnerProcesses() {
     try {
-      this.energyStore.utilize(INNER_PROCESS_ENERGY_RATE);
+      this.energyStore.utilize(this.parameters.innerProcessEnergyCost);
     } catch (e: unknown) {
       if (e instanceof NotEnoughEnergyError) {
         // consider changing the reproduction strategy
@@ -68,7 +90,7 @@ export class Cell {
   }
 
   private consume() {
-    const consumed = this.terrain.consume(RESOURCE_INTAKE);
+    const consumed = this.terrain.consume(this.parameters.resourceConsumptionRate);
 
     if (consumed === 0) {
       this.move();
@@ -80,8 +102,8 @@ export class Cell {
 
   private grow() {
     try {
-      this.energyStore.utilize(GROW_PROCESS_ENERGY_RATE);
-      this.size += GROW_PROCESS_ENERGY_RATE;
+      this.energyStore.utilize(this.parameters.growEnergyCost);
+      this.size += this.parameters.growRate;
     } catch (e: unknown) {
       if (e instanceof NotEnoughEnergyError) {
         // consider changing the reproduction strategy
@@ -100,7 +122,7 @@ export class Cell {
     }
 
     const clone = this.clone();
-    this.energyStore.utilize(DIVISION_PROCESS_ENERGY_RATE);
+    this.energyStore.utilize(this.parameters.divisionEnergyCost);
     this.terrain.put(clone, randomArrayElement(freeSpots));
     this.children.push(clone);
   }
@@ -113,7 +135,7 @@ export class Cell {
 
     const nextMove = randomArrayElement(nextMoves);
     try {
-      this.energyStore.utilize(MIGRATION_PROCESS_ENERGY_RATE);
+      this.energyStore.utilize(this.parameters.migrationEnergyCost);
       this.terrain.moveToSpot(nextMove);
     } catch (e: unknown) {
       if (e instanceof NotEnoughEnergyError) {
@@ -128,7 +150,6 @@ export class Cell {
     do {
       id = `${rootAncestorId}_${randomStr(4)}`;
     } while (this.children.find((cell) => cell.id === id) !== undefined);
-    const cell = new Cell(id, this.terrain.clone(id));
-    return cell;
+    return new Cell(id, this.terrain.clone(id), this.parameters);
   }
 }