+++ src/main.ts
@@ -1,10 +1,9 @@
 export { Cell } from "./cultures/Cell";
-export { CellFactory, type CellConfig } from "./cultures/CellFactory";
+export { CellFactory } from "./cultures/CellFactory";
+export { type CellParameters } from "./cultures/Cell";
 
 export { Terrain } from "./terrain/Terrain";
 export { TerrainFactory, type TerrainConfig } from "./terrain/TerrainFactory";
 
 export { Visualizer } from "./visualizer/Visualizer";
 export { Simulation } from "./Simulation";
-
-export { generateIds } from "./util/id/generator";